refactorerd and created profiles controller, also included the routing in routes.rb from line 33-36
