+++ app/views/view_purse/view_abilities.html.erb
@@ -1,0 +1,2 @@
+<h1>ViewPurse#view_abilities</h1>
+<p>Find me in app/views/view_purse/view_abilities.html.erb</p>
