+++ app/views/view_purse/view_stats.html.erb
@@ -1,0 +1,2 @@
+<h1>ViewPurse#view_stats</h1>
+<p>Find me in app/views/view_purse/view_stats.html.erb</p>
